refactor(materials/vue): restructure materials components and update imports
- Updated the structure of materials components by creating a new 'materials' directory and moving relevant files.
- Introduced new components such as ActionButton, GridStack, and GridStackItem.
- Updated package.json and vite.config.ts to reflect the new materials structure.
- Adjusted imports in playground files to use the new materials path.
- Added deprecation notice for old glob versions in pnpm-lock.yaml.

This change enhances the organization of materials components and improves the overall maintainability of the codebase.
