--- packages/materials/vue-opentiny-vue/src/components/index.ts
@@ -1,1 +1,0 @@
-export { vueMaterials } from './vue-materials';
