+++ packages/materials/vue-opentiny-vue/src/materials/components/GridStackItem.vue
