+++ packages/materials/vue-opentiny-vue/src/materials/components/TinyTabsWrap.vue
