+++ packages/materials/vue-opentiny-vue/src/materials/index.ts
@@ -1,0 +1,2 @@
+export * from './materials';
+export * from './components';
