+++ packages/materials/vue-opentiny-vue/src/materials/materials.ts
@@ -1,0 +1,9 @@
+import { components, type IComponents } from './components';
+
+export interface IMaterials {
+  components: IComponents;
+}
+
+export const materials: IMaterials = {
+  components,
+};
