+++ packages/materials/vue-opentiny-vue/vite.config.ts
@@ -19,7 +19,7 @@ export default defineConfig({
       entry: {
         index: path.resolve(__dirname, './src/index.ts'),
         'render-config': path.resolve(__dirname, './src/render-config/index.ts'),
-        'components': path.resolve(__dirname, './src/components/index.ts'),
+        'materials': path.resolve(__dirname, './src/materials/index.ts'),
       },
       formats: ['es'],
       fileName: (format, entryName) => `${entryName}.js`,
