+++ sites/playground/web/src/App.vue
@@ -14,7 +14,7 @@ import {
   h,
   shallowRef,
 } from 'vue';
-import { vueMaterials } from '@opentiny/genui-sdk-materials-vue-opentiny-vue/components';
+import { materials } from '@opentiny/genui-sdk-materials-vue-opentiny-vue/materials';
 import { getModelFeatures, getModelOptions } from './api';
 import { createCustomFetch } from './api/custom-fetch';
 import AssistantFooter from './components/AssistantFooter.vue';
@@ -368,7 +368,7 @@ onUnmounted(() => {
         </div>
       </template>
       <div v-show="!ENABLE_TEMPLATE || activeName !== 'template'" class="chat-container">
-        <GenuiConfigProvider :theme="theme" :locale="locale" :materials="{ ...vueMaterials }" style="height: 100%">
+        <GenuiConfigProvider :theme="theme" :locale="locale" :materials="materials.components" style="height: 100%">
           <GenuiChat
             :url="url"
             ref="chat"
