+++ sites/playground/web/tsconfig.dev.json
@@ -21,8 +21,8 @@
       "@opentiny/genui-sdk-materials-vue-opentiny-vue/render-config": [
         "../../../packages/materials/vue-opentiny-vue/src/render-config/index.ts"
       ],
-      "@opentiny/genui-sdk-materials-vue-opentiny-vue/components": [
-        "../../../packages/materials/vue-opentiny-vue/src/components/index.ts"
+      "@opentiny/genui-sdk-materials-vue-opentiny-vue/materials": [
+        "../../../packages/materials/vue-opentiny-vue/src/materials/index.ts"
       ],
       "@opentiny/genui-sdk-materials-angular-opentiny-ng": [
         "../../../packages/materials/angular-opentiny-ng/src/index.ts"
